requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
pyrate-limiter==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-ratelimiter==0.7.0
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
